split runner_test.go
